Rename custom thread type.
On mingw, using the name 'thread', even though it is inside a namespace,
causes it to behave "normally". I don't know exactly why, but I'm guessing
that if the name and the implementation match, something lines up well
enough in the linking stage and this custom thread implementation ends up
using the standard library implementation.

This was noticed when trying to compile on linux. The linux compiler or
the standard library apparently don't match things up in the same way, so
the implementation broke. This was expected, but it also revealed that it
should have been broken on mingw as well, since there was no custom
implementation included in the code.